fix ci error
